requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.16.1
anyio==3.2.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2022.5.18.1
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
fastapi-login==1.6.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
iso8601==0.1.14
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
passlib==1.7.4
pydantic==1.8.1
PyJWT==2.1.0
pypika-tortoise==0.1.5
python-multipart==0.0.5
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.5.0
secure==0.3.0
six==1.16.0
slack-sdk==3.4.2
sniffio==1.2.0
starlette==0.13.6
tortoise-orm==0.17.3
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
yarl==1.6.3
